now only accepts verified models and adds trigger word to prompt automatically

+++ hordecommands.py
@@ -3,10 +3,13 @@ import json
 
 async def setmodelglobal(arg: list) -> str:
     arg = (" ".join(arg)).strip()
+    with open("acceptedmodels.json", "r+") as file:
+        models = json.load(file)
     with open("settings.json", "r+") as file:
         data = json.load(file)
-        if(arg in data['models']):
-            data['defaultpayload']['models'] = [arg]
+        if(arg in data['models'] and arg in list(models)):
+            data['defaultpayload']['models'] = [models[arg]['modelname']]
+            data['keyword'] = models[arg]['Keyword']
             file.seek(0)
             json.dump(data, file, indent=4)
             file.truncate()
@@ -28,3 +31,11 @@ def get_website()->str:
     with open("settings.json", "r+") as file:
         data = json.load(file)
     return data["website"]
+
+def info_model(arg: str) -> str:
+    with open("acceptedmodels.json", 'r') as file:
+        data = json.load(file)
+    if(arg in list(data)):
+        return f"> Modelname: {data[arg]['modelname']}\n> Description: {data[arg]['Description']}\n> Website: {data[arg]['Website']}\n> Keyword: '{data[arg]['Keyword']}'"
+    return "No model with that tag."
+
